improving the way the snake extends
In this way the new segment is added at the end of the snake. This makes
easier to play the game.
When the segment is added at the head the snake moves 2 times and when
the food is near to the border or the tail is more probably to lose